# Conflicts:
#	Código Fonte/VERSAO-ATUAL/src/Credenciado/Credenciado/Web.config
#	Código Fonte/VERSAO-ATUAL/src/EtramiteX/Interno/Areas/Caracterizacoes/Views/RegularizacaoFundiaria/DadosDominialidadePartial.ascx
#	Código Fonte/VERSAO-ATUAL/src/EtramiteX/Model.Extencao/Caracterizacoes/ModuloRegularizacaoFundiaria/Data/RegularizacaoFundiariaDa.cs

+++ Código Fonte/VERSAO-ATUAL/lib/src/Blocos.Entities/Interno.Extensoes/Caracterizacoes/ModuloRegularizacaoFundiaria/Posse.cs
@@ -26,7 +26,7 @@ namespace Tecnomapas.Blocos.Entities.Interno.Extensoes.Caracterizacoes.ModuloReg
 	
 		public Decimal AreaPosseDocumento { get; set; }
 		public String DescricaoComprovacao { get; set; }
-		public Int32? NumeroCCIR { get; set; }
+		public long? NumeroCCIR { get; set; }
 		public Decimal AreaCCIR { get; set; }
 
 		private DateTecno _dataUltimaAtualizacaoCCIR = new DateTecno();
